moved Contact folder
from RadicalCollaboration to pages
