+++ src/App.js
@@ -4,7 +4,7 @@ import { QueryParamProvider } from 'use-query-params';
 import Layout from './components/common/Layout';
 import ScrollToTop from './components/common/ScrollToTop';
 import About from './pages/About';
-import Contact from './pages/RadicalCollaboration/Contact';
+import Contact from './pages/Contact';
 import Collaborate from './pages/Collaborate';
 import Contributors from './pages/Contributors';
 import Donation from './pages/Donation';
