+++ src/pages/Contact/ContactForm.js
@@ -6,7 +6,7 @@ import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
-import FileUploadIcon from '../../../icons/FileUploadIcon';
+import FileUploadIcon from '../../icons/FileUploadIcon';
 
 const useStyles = makeStyles((theme) => ({
   formStyle: {
