+++ src/pages/Contact/index.js
@@ -5,7 +5,7 @@ import makeStyles from '@material-ui/core/styles/makeStyles';
 import Typography from '@material-ui/core/Typography';
 import Card from '@material-ui/core/Card';
 import CardMedia from '@material-ui/core/CardMedia';
-import { GenericHeaderSection } from '../../../components/';
+import { GenericHeaderSection } from '../../components/';
 
 const useStyles = makeStyles((theme) => ({
   headerStyle: {
